+++ python/GafferSceneTest/StandardAttributesTest.py
@@ -169,7 +169,7 @@ class StandardAttributesTest( GafferSceneTest.SceneTestCase ) :
 
 		targets = "attribute:scene:visible attribute:doubleSided attribute:render:* attribute:gaffer:* " \
 			+ "attribute:linkedLights attribute:linkedLights:exclusions attribute:shadowedLights " \
-			+ "attribute:shadowedLights:exclusions attribute:filteredLights"
+			+ "attribute:shadowedLights:exclusions attribute:filteredLights attribute:filteredLights:exclusions"
 
 		node = GafferScene.StandardAttributes()
 		for attribute in Gaffer.Metadata.targetsWithMetadata( targets, "defaultValue" ) :